OglTexture: fixing MAX_NUMBER_OF_TEXTURE_UNIT for some nvidia cards
Former-commit-id: 637f57fc602eee4084a70bfa69a58ac4c699ec77

+++ modules/SofaOpenglVisual/OglTexture.h
@@ -82,7 +82,7 @@ protected:
     helper::io::Image* img;
 
 public:
-    static unsigned short MAX_NUMBER_OF_TEXTURE_UNIT;
+    static GLint MAX_NUMBER_OF_TEXTURE_UNIT;
 protected:
     OglTexture();
     virtual ~OglTexture();
